4.0.0 upgrade to python3

+++ code/default/gae_proxy/server/gae/legacy.py
@@ -9,6 +9,6 @@ def application(environ, start_response):
     if environ['PATH_INFO'] == '/robots.txt':
         yield '\n'.join(['User-agent: *', 'Disallow: /'])
     else:
-        timestamp = long(environ['CURRENT_VERSION_ID'].split('.')[1])/2**28
+        timestamp = int(environ['CURRENT_VERSION_ID'].split('.')[1])/2**28
         ctime = time.strftime('%Y-%m-%d %H:%M:%S', time.gmtime(timestamp+8*3600))
         yield "GoAgent 服务端已经在 %s 升级到 %s 版本, 请更新您的客户端。" % (ctime, __version__)
